Bugfix/general

+++ backend/FinanceApp.Application/DTOs/Goal/CreateGoalDto.cs
@@ -9,6 +9,6 @@ public class CreateGoalDto
 {
     public string? Name { get; set; }
     public string Category { get; set; } = null!;
-    public int TargetAmount { get; set; }
+    public decimal TargetAmount { get; set; }
     public DateTime? DesiredFinish { get; set; }
 }
