+++ backend/FinanceApp.Application/DTOs/Goal/GoalDto.cs
@@ -10,8 +10,8 @@ public class GoalDto
     public int Id { get; set; }
     public string? Name { get; set; }
     public string Category { get; set; } = null!;
-    public int SavedAmount { get; set; }
-    public int TargetAmount { get; set; }
+    public decimal SavedAmount { get; set; }
+    public decimal TargetAmount { get; set; }
     public DateTime StartTime { get; set; }
     public DateTime? DesiredFinish { get; set; }
 }
