+++ backend/FinanceApp.Application/DTOs/Transaction/TransactionDto.cs
@@ -11,6 +11,6 @@ public class TransactionDto
     public string? Name { get; set; }
     public string Category { get; set; } = null!;
     public string? Description { get; set; }
-    public int Amount { get; set; }
+    public decimal Amount { get; set; }
     public DateTime Time { get; set; }
 }
